requirements.txt
flask
flask-cors
python-dotenv
Werkzeug==2.2.2
google-generativeai
PyPDF2